+++ core/vm/contracts.go
@@ -49,8 +49,8 @@ type PrecompiledContract interface {
 // PrecompiledContractWithChainConfig is an interface that defines a method to set the
 // chain config for precompiled contract.
 type PrecompiledContractWithChainConfig interface {
-	// SetChainConfig sets isHomestead, isIstanbul and isShanghai config for precompiled contract.
-	SetChainConfig(isHomestead, isIstanbul, isShanghai bool)
+	// WithChainConfig copies isHomestead, isIstanbul and isShanghai config for precompiled contract.
+	WithChainConfig(isHomestead, isIstanbul, isShanghai bool) PrecompiledContract
 }
 
 // PrecompiledContractsHomestead contains the default set of pre-compiled Ethereum
@@ -275,7 +275,7 @@ func (evm *EVM) RunPrecompiledContract(
 	readOnly bool,
 ) (ret []byte, remainingGas uint64, err error) {
 	if pp, ok := p.(PrecompiledContractWithChainConfig); ok {
-		pp.SetChainConfig(evm.chainRules.IsHomestead, evm.chainRules.IsIstanbul, evm.chainRules.IsShanghai)
+		p = pp.WithChainConfig(evm.chainRules.IsHomestead, evm.chainRules.IsIstanbul, evm.chainRules.IsShanghai)
 	}
 	return runPrecompiledContract(evm, p, caller, input, suppliedGas, value, readOnly)
 }